+++ src/ui/common/src/utils/shared.ts
@@ -1,3 +1,6 @@
+import { TableRow } from './data';
+import { ArtifactResultMetadataResponse } from "src/handlers/responses/artifact";
+
 export enum LoadingStatusEnum {
   Initial = 'initial',
   Loading = 'loading',
@@ -36,6 +39,29 @@ export enum ExecutionStatus {
   Running = 'running',
   // Checks can have a warning status.
   Warning = 'warning',
+}
+
+export const getArtifactExecStateAsTableRow = (artifactStatusResult: ArtifactResultMetadataResponse): TableRow => {
+  const all_times = [
+    artifactStatusResult.exec_state?.timestamps?.finished_at,
+    artifactStatusResult.exec_state?.timestamps?.pending_at,
+    artifactStatusResult.exec_state?.timestamps?.registered_at,
+    artifactStatusResult.exec_state?.timestamps?.running_at,
+  ];
+
+  const times = all_times.filter((x) => typeof x === 'string').map((x) => new Date(x)); // Convert from string to time
+
+  const maxTime = Math.max.apply(null, times); // Returns -Infinity if times is an empty list
+  
+  // Need to convert back to Date because math.max changes the time to Unix time (number)
+  const timestamp = 
+    maxTime > 0? new Date(maxTime).toLocaleString() : 'Unknown'; 
+    
+  return {
+    timestamp,
+    status: artifactStatusResult.exec_state?.status ?? 'Unknown',
+    value: artifactStatusResult.content_serialized,
+  }
 }
 
 export const stringToExecutionStatus = (status: string): ExecutionStatus => {
